backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
Pillow==10.1.0
python-dotenv==1.0.0
drf-yasg==1.21.7
pyjwt==2.8.0